tests/wpt/web-platform-tests/eventsource/format-field-retry-empty.htm

<!doctype html>
<html>
  <head>
    <title>EventSource: empty retry field</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var source = new EventSource("resources/message.py?message=retry%0Adata%3Atest")
        source.onmessage = function(e) {
          test.step(function() {
            assert_equals("test", e.data)
            source.close()
          })
          test.done()
        }
      })
    </script>
  </body>
</html>
